+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue where clicking into Constant name field and leaving without changing anything would cause compiled symbol to change
 - Fix issue where changing a Constant's name and clicking away could sometimes cause name change to be lost
 - Fix issue where renaming a variable to have the same name but with different capitalisation could cause compiled symbol to change
+- Fix issue where debugger wouldn't display prefab scripts
 
 ## [4.3.2] - 2026-06-22